update README.md and replace @AutoConfiguration
