+++ community/plugins/spring-ai-alibaba-starter-plugin-baidusearch/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.alibaba.cloud.ai.plugin.baidu.BaiduSearchPluginConfiguration
+com.alibaba.cloud.ai.plugin.baidu.BaiduSearchPluginAutoConfiguration
