+++ community/plugins/spring-ai-alibaba-starter-plugin-bingsearch/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,2 @@
-com.alibaba.cloud.ai.plugin.bing.BingSearchPluginConfiguration
+com.alibaba.cloud.ai.plugin.bing.BingSearchPluginAutoConfiguration
