+++ community/plugins/spring-ai-alibaba-starter-plugin-dingtalk/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.alibaba.cloud.ai.plugin.dingtalk.DingTalkConfiguration
+com.alibaba.cloud.ai.plugin.dingtalk.DingTalkAutoConfiguration
