+++ community/plugins/spring-ai-alibaba-starter-plugin-gaode/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.alibaba.cloud.ai.plugin.gaode.GaoDeConfiguration
+com.alibaba.cloud.ai.plugin.gaode.GaoDeAutoConfiguration
